Dejamos solo la pantalla principal y le quitamos el TopBar, dejando solo el buscador.
