+++ app/src/main/java/com/example/megagifs/principalscreen/ui/BottomNavigationPrincipal.kt
@@ -1,0 +1,61 @@
+package com.example.megagifs.principalscreen.ui
+
+import androidx.compose.material.BottomNavigation
+import androidx.compose.material.BottomNavigationItem
+import androidx.compose.material.Icon
+import androidx.compose.material.Text
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Favorite
+import androidx.compose.material.icons.filled.Home
+import androidx.compose.material.icons.filled.Person
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+
+/*****
+ * Proyect: MegaGifs
+ * Package: com.example.megagifs.screen3.ui
+ *
+ * Created by Rafael Barbeyto Torrellas on 02/06/2023 at 13:52
+ * More info: https://www.linkedin.com/in/rafa-barbeyto/
+ *
+ * All rights reserved 2023.
+ *****/
+
+@Composable
+fun MyBottomNavigation() {
+
+    var index by rememberSaveable { mutableStateOf(0) }
+
+    BottomNavigation(
+        elevation = 24.dp,
+        backgroundColor = Color.DarkGray,
+        contentColor = Color.Yellow
+    ) {
+        BottomNavigationItem(selected = index == 0, onClick = { index = 0 }, icon = {
+            Icon(
+                imageVector = Icons.Default.Home,
+                contentDescription = "Home",
+                tint = Color.Yellow
+            )
+        }, label = { Text(text = "Home") })
+        BottomNavigationItem(selected = index == 1, onClick = { index = 1 }, icon = {
+            Icon(
+                imageVector = Icons.Default.Favorite,
+                contentDescription = "Fav",
+                tint = Color.Yellow
+            )
+        }, label = { Text(text = "Fav") })
+        BottomNavigationItem(selected = index == 2, onClick = { index = 2 }, icon = {
+            Icon(
+                imageVector = Icons.Default.Person,
+                contentDescription = "Person",
+                tint = Color.Yellow
+            )
+        }, label = { Text(text = "Person") })
+    }
+}
